pyproject.toml
[tool.poetry]
name = "crypto-data-platform"
version = "0.1.0"
description = "A data platform for crypto trading analytics with ETL pipeline and dashboard"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
requests = "^2.31.0"
pandas = "^2.1.0"
sqlalchemy = "^2.0.20"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"
pyyaml = "^6.0.1"
loguru = "^0.7.2"
python-dateutil = "^2.8.2"
pytz = "^2023.3"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
pydantic = "^2.6.3"
aiofiles = "^23.2.1"
jinja2 = "^3.1.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.7.0"
isort = "^5.12.0"
mypy = "^1.5.1"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=src --cov-report=term-missing" 